+++ lib/universal_feed.dart
@@ -29,6 +29,7 @@ export 'src/universal/extensions/podcast/podcast_chapters.dart';
 export 'src/universal/extensions/podcast/podcast_item.dart';
 export 'src/universal/extensions/podcast/podcast_transcript.dart';
 export 'src/universal/extensions/podcast/podcast_txt.dart';
+export 'src/universal/extensions/podcast/podcast_update_frequency.dart';
 export 'src/universal/generator.dart';
 export 'src/universal/image.dart';
 export 'src/universal/item.dart';